requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.3
exceptiongroup==1.2.0
executing==2.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.0
pyOpenSSL==23.3.0
PyYAML==6.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.6.0
template==0.7.6
toml==0.10.2
traitlets==5.13.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
zope.interface==6.1
